requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
dj-static==0.0.6
Django==5.2.11
django-ckeditor-5==0.2.20
django-colorfield==0.14.0
django-js-asset==3.1.2
idna==3.11
pillow==12.1.1
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
static3==0.7.0
tzdata==2025.3
urllib3==2.6.3
